chore: remove slimrpc from bindings
