+++ crates/rpc/Cargo.toml
@@ -2,17 +2,25 @@
 # SPDX-License-Identifier: Apache-2.0
 
 [package]
-name = "agntcy-slim-rpc"
+name = "agntcy-slimrpc"
 edition = { workspace = true }
 license = { workspace = true }
 version = "0.1.0"
 description = "SlimRPC: a gRPC-like RPC framework over SLIM (native Rust)."
 
 [lib]
-name = "slim_rpc"
+name = "slimrpc"
+
+[features]
+default = []
+# Enables the UniFFI foreign-language interface. This pulls in `agntcy-slim-bindings`
+# for the `App`/`Name` wrapper types exposed across the FFI boundary and emits the
+# UniFFI scaffolding. Without this feature the crate is a pure native Rust library.
+bindings = ["dep:uniffi", "dep:agntcy-slim-bindings"]
 
 [dependencies]
 agntcy-slim-auth = { workspace = true }
+agntcy-slim-bindings = { workspace = true, optional = true }
 agntcy-slim-datapath = { workspace = true }
 agntcy-slim-service = { workspace = true, features = ["session"] }
 agntcy-slim-session = { workspace = true }
@@ -27,6 +35,7 @@ thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
 tracing = { workspace = true }
+uniffi = { workspace = true, features = ["cli"], optional = true }
 uuid = { workspace = true }
 
 [dev-dependencies]
