+++ crates/rpc/src/stream_types.rs
@@ -7,6 +7,7 @@
 //! Since UniFFI doesn't support Rust async streams, we provide synchronous
 //! pull/push interfaces backed by async channels.
 
+#[cfg(not(feature = "bindings"))]
 use slim_datapath::api::ProtoName as Name;
 use std::pin::Pin;
 use std::task::Poll;
@@ -160,6 +161,7 @@ impl StreamSource {
 /// Allows pulling messages from a client request stream.
 /// This wraps the underlying async stream and provides a blocking interface
 /// suitable for UniFFI callback traits.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct RequestStream {
     /// Inner stream wrapped in a mutex for interior mutability
     inner: TokioMutex<DecodedStream<Vec<u8>>>,
@@ -174,6 +176,7 @@ impl RequestStream {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl RequestStream {
     /// Pull the next message from the stream (blocking version)
     ///
@@ -196,6 +199,7 @@ impl RequestStream {
 }
 
 /// Message from a stream
+#[cfg_attr(feature = "bindings", derive(uniffi::Enum))]
 pub enum StreamMessage {
     /// Successfully received data
     Data(Vec<u8>),
@@ -210,6 +214,7 @@ pub enum StreamMessage {
 /// Allows pushing messages to a client response stream.
 /// This wraps an async channel sender and provides a blocking interface
 /// suitable for UniFFI callback traits.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct ResponseSink {
     /// Channel sender for streaming responses (None when closed)
     sender: Mutex<Option<RpcBytesSender>>,
@@ -231,6 +236,7 @@ impl ResponseSink {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl ResponseSink {
     /// Send a message to the response stream (blocking version)
     ///
@@ -310,6 +316,7 @@ impl ResponseSink {
 /// Response stream reader for unary-to-stream RPC calls
 ///
 /// Allows pulling messages from a server response stream one at a time.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct ResponseStreamReader {
     /// Inner receiver channel for stream messages
     inner: TokioMutex<UnboundedReceiver<Result<Vec<u8>, RpcError>>>,
@@ -324,6 +331,7 @@ impl ResponseStreamReader {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl ResponseStreamReader {
     /// Pull the next message from the response stream (blocking version)
     ///
@@ -348,6 +356,7 @@ impl ResponseStreamReader {
 /// Request stream writer for stream-to-unary RPC calls
 ///
 /// Allows sending multiple request messages and getting a final response.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct RequestStreamWriter {
     sender: TokioMutex<Option<UnboundedSender<Vec<u8>>>>,
     response: TokioMutex<Option<RpcBytesJoinHandle>>,
@@ -395,6 +404,7 @@ impl RequestStreamWriter {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl RequestStreamWriter {
     /// Send a request message to the stream (blocking version)
     pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
@@ -443,6 +453,7 @@ impl RequestStreamWriter {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl RequestStreamWriter {
     /// Finalize the stream and get the response (async version)
     ///
@@ -455,6 +466,7 @@ impl RequestStreamWriter {
 /// Bidirectional stream handler for stream-to-stream RPC calls
 ///
 /// Allows sending and receiving messages concurrently.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct BidiStreamHandler {
     sender: TokioMutex<Option<UnboundedSender<Vec<u8>>>>,
     receiver: TokioMutex<UnboundedReceiver<Result<Vec<u8>, RpcError>>>,
@@ -504,6 +516,7 @@ impl BidiStreamHandler {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl BidiStreamHandler {
     /// Send a request message to the stream (blocking version)
     pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
@@ -557,14 +570,20 @@ impl BidiStreamHandler {
 /// Per-message context for a multicast RPC response — identifies which group
 /// member sent the response.
 #[derive(Clone, Debug)]
+#[cfg_attr(feature = "bindings", derive(uniffi::Record))]
 pub struct RpcMessageContext {
     /// The SLIM name of the group member that sent this response.
+    #[cfg(not(feature = "bindings"))]
     pub source: Arc<Name>,
+    /// The SLIM name of the group member that sent this response.
+    #[cfg(feature = "bindings")]
+    pub source: Arc<slim_bindings::Name>,
 }
 
 /// A single item in a multicast response stream, pairing the response payload
 /// with the identity of the member that produced it.
 #[derive(Clone, Debug)]
+#[cfg_attr(feature = "bindings", derive(uniffi::Record))]
 pub struct RpcMulticastItem {
     /// Context identifying the source member.
     pub context: RpcMessageContext,
@@ -573,6 +592,7 @@ pub struct RpcMulticastItem {
 }
 
 /// Message from a multicast response stream.
+#[cfg_attr(feature = "bindings", derive(uniffi::Enum))]
 pub enum MulticastStreamMessage {
     /// Successfully received response item with source context.
     Data { item: RpcMulticastItem },
@@ -586,6 +606,7 @@ pub enum MulticastStreamMessage {
 ///
 /// Allows pulling `RpcMulticastItem`s from a GROUP response stream one at a
 /// time. Each item carries the source member's identity alongside the payload.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct MulticastResponseReader {
     inner: TokioMutex<UnboundedReceiver<Result<MulticastItem<Vec<u8>>, RpcError>>>,
 }
@@ -603,19 +624,24 @@ impl MulticastResponseReader {
         item: Result<MulticastItem<Vec<u8>>, RpcError>,
     ) -> MulticastStreamMessage {
         match item {
-            Ok(mi) => MulticastStreamMessage::Data {
-                item: RpcMulticastItem {
-                    context: RpcMessageContext {
-                        source: Arc::new(mi.context.source),
+            Ok(mi) => {
+                #[cfg(not(feature = "bindings"))]
+                let source = Arc::new(mi.context.source);
+                #[cfg(feature = "bindings")]
+                let source = Arc::new(slim_bindings::Name::from_slim_name(mi.context.source));
+                MulticastStreamMessage::Data {
+                    item: RpcMulticastItem {
+                        context: RpcMessageContext { source },
+                        message: mi.message,
                     },
-                    message: mi.message,
-                },
-            },
+                }
+            }
             Err(e) => MulticastStreamMessage::Error { error: e },
         }
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl MulticastResponseReader {
     /// Pull the next item from the multicast response stream (blocking).
     pub fn next(&self) -> MulticastStreamMessage {
@@ -639,6 +665,7 @@ impl MulticastResponseReader {
 /// close the request stream via [`close_send`](Self::close_send), and receive
 /// responses via [`recv`](Self::recv) / [`recv_async`](Self::recv_async). Each
 /// response item carries the source member's identity.
+#[cfg_attr(feature = "bindings", derive(uniffi::Object))]
 pub struct MulticastBidiStreamHandler {
     sender: TokioMutex<Option<UnboundedSender<Vec<u8>>>>,
     receiver: TokioMutex<UnboundedReceiver<Result<MulticastItem<Vec<u8>>, RpcError>>>,
@@ -689,6 +716,7 @@ impl MulticastBidiStreamHandler {
     }
 }
 
+#[cfg_attr(feature = "bindings", uniffi::export)]
 impl MulticastBidiStreamHandler {
     /// Send a request message to the stream (blocking).
     pub fn send(&self, data: Vec<u8>) -> Result<(), RpcError> {
